refactor: post.tsxを３つのコンポーネントに切り出し_#76
